+++ data-raw/cow_mid_ddydisps.R
@@ -92,5 +92,7 @@ cow_mid_ddydisps %>%
   rename(cowmidongoing = dispongoing,
          cowmidonset = disponset) -> cow_mid_ddydisps
 
+
+
 save(cow_mid_ddydisps, file="data/cow_mid_ddydisps.rda")